+++ core/modules/tests.rs
@@ -409,7 +409,7 @@ fn test_mods() {
 }
 
 #[test]
-fn test_mods_() {
+fn test_lazy_loaded_esm() {
   deno_core::extension!(test_ext, lazy_loaded_esm = [dir "modules/testdata", "lazy_loaded.js"]);
 
   let mut runtime = JsRuntime::new(RuntimeOptions {
@@ -425,6 +425,8 @@ fn test_mods_() {
       const module = Deno.core.ops.op_lazy_load_esm("ext:test_ext/lazy_loaded.js");
       module.blah("hello\n");
       Deno.core.print(`${JSON.stringify(module)}\n`);
+      const module1 = Deno.core.ops.op_lazy_load_esm("ext:test_ext/lazy_loaded.js");
+      if (module !== module1) throw new Error("should return the same error");
       "#,
     )
     .unwrap();
